+++ base64encode.h
@@ -11,6 +11,7 @@ For details, see http://sourceforge.net/projects/libb64
 #ifndef BASE64_ENCODE_H
 #define BASE64_ENCODE_H
 
+#include <stddef.h>
 #include <stdint.h>
 
 typedef enum {
@@ -24,8 +25,8 @@ typedef struct {
 
 void base64_encode_init(base64_encodestate *);
 char base64_encode_value(uint8_t);
-int base64_encode_update(base64_encodestate *, const uint8_t *, uint64_t, char *);
-int base64_encode_final(base64_encodestate *, char *);
-int base64_encode(const uint8_t *, uint64_t, char *);
+ptrdiff_t base64_encode_update(base64_encodestate *, const uint8_t *, uint64_t, char *);
+ptrdiff_t base64_encode_final(base64_encodestate *, char *);
+ptrdiff_t base64_encode(const uint8_t *, uint64_t, char *);
 
 #endif /* BASE64_ENCODE_H */
